adicionando máscaras e validações de campos

+++ ProjetoAW/Models/ItemPedido.cs
@@ -5,6 +5,6 @@
         public int codItemPedido { get; set; }
         public string nomeItemPedido { get; set; }
         public int quantidadeItem { get; set; }
-        public double valorParcial { get; set; }
+        public decimal valorParcial { get; set; }
     }
 }
